reordenar y simplificar codigo
